requirements.txt
django
tensorflow-macos
opencv-python
numpy
matplotlib
pillow
